[jsk_recognition_utils] Separate segment.h from geo_util.h
